Clean up project after implementation of page factory
